arm64: dts: qcom: msm8916-samsung-a5-zt: Add initial device tree
Samsung Galaxy A5 (SM-A500YZ) is a smartphone using the MSM8916 SoC released
in 2015.

The A5 variants are very similar, with some differences in accelerometer and
touch key. The common parts are shared in msm8916-samsung-a5u-eur.dts to reduce
duplication.

The difference of accelerometer between A5U/A5ZT is the way the sensor is
mounted on the mainboard - set the mount-matrix in the
device-specific device tree part to handle that difference.

On A5ZT the touch key is supplied by a single fixed
regulator (enabled via GPIO 60) that supplies both MCU and LED.

A5ZT doesn't have an NFC chip.

Signed-off-by: Lin, Meng-Bo <linmengbo0689@protonmail.com>