Input: cyttsp4 - Import forward ported downstream driver
This is a minimized import of the cyttsp4 touchscreen driver from
Samsung's downstream G357FZXXU1AOE1 heatqlte kernel, with forward
porting to get it compiling on mainline Linux, mainly from simply going
through the git history and applying changes previously made for the
original mainline cyttsp4 driver, and with changes to clean it up in
general and remove some unneeded functionality.

The mainline driver was added by Cypress all the way back in Linux v3.11
and then unfortunately shortly after abandoned. It's missing support for
device tree configuration, and maybe other basic things, and is
apparently buggy. This imported downstream version of the driver is more
mature, but also in a not too great messy state typical for a downstream
driver.

Seeing that the work had been done on maturing the driver, surely
Cypress could've cleaned up and sent more patches upstream to bring the
mainline version into a good state. But unfortunately I guess that never
happened for whatever reason, and won't.

Meanwhile, it would be a real challenge for someone to improve the
existing mainline driver, and take things from a downstream version.
Perhaps, in all likelihood, that will all never happen, and so using
this downstream driver it shall be.

Thankfully, for long term usage if need be, applying changes to keep the
driver compiling and working shouldn't be a problem, since pretty much
just applying changes from commits applied to the mainline driver was
sufficient to get it compiling, and then working without issues.

Small issue that could be resolved - Figure out how this is occurring:
[    2.341564] cyttsp4_i2c_adapter 4-0024: Runtime PM usage count underflow!

Also, all parts of the driver are compiled into a single module, even
though there is meant to be more separation between the parts. But they
don't compile into individual modules successfully, the downstream
kernel has them setup as individual objects which would similarly also
cause problems if compiling the driver as a module there, and getting
them to be separate looks tricky and not worthwhile for a downstream
driver.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>