arm64: dts: qcom: msm8916-samsung-heatqlte: Add touchkeys
For its touchkeys, the Samsung Galaxy Ace 4 uses a Cypress CY8CMBR3145
touchkey controller, which is communicated with using software
bit-banging through i2c-gpio, and uses two separate regulators for the
touchkey input and LEDs.

It works fine with the TM2 touchkey driver and its TM2 configuration,
both with input and with the LEDs working. Add device tree configuration
for it.

Worth noting that despite `fixed_regulator` not being set to true for
the TM2 configuration, and fixed regulators are being used, there aren't
any problems. With it set to true, the only apparent difference in
behaviour is that the driver enables the LEDs for a brief second on boot
before disabling them.

Signed-off-by: K9100ii <k9100ii.ki@gmail.com>